src/main/kotlin/com/aheidelbacher/algoventure/core/input/Input.kt
/*
 * Copyright 2016 Andrei Heidelbacher <andrei.heidelbacher@gmail.com>
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.aheidelbacher.algoventure.core.input

import com.fasterxml.jackson.annotation.JsonSubTypes
import com.fasterxml.jackson.annotation.JsonTypeInfo

@JsonTypeInfo(
        use = JsonTypeInfo.Id.NAME,
        include = JsonTypeInfo.As.PROPERTY,
        property = "type"
)
@JsonSubTypes(
        JsonSubTypes.Type(value = Input.Click::class, name = "Click"),
        JsonSubTypes.Type(value = Input.Scroll::class, name = "Scroll"),
        JsonSubTypes.Type(value = Input.Wait::class, name = "Wait")
)
sealed class Input {
    companion object {
        const val INPUT: String = "input"
    }

    class Click(val x: Int, val y: Int): Input()

    class Scroll(val dx: Int, val dy: Int): Input()

    object Wait : Input()
}
